src/test/java/io/github/oclay1st/wfdb/mocks/formatters/SignalFormatter61Test.java
package io.github.oclay1st.wfdb.mocks.formatters;

import static org.junit.jupiter.api.Assertions.assertArrayEquals;
import static org.junit.jupiter.api.Assertions.assertNotNull;

import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import io.github.oclay1st.wfdb.HeaderSignal;
import io.github.oclay1st.wfdb.SignalFormat;
import io.github.oclay1st.wfdb.formatters.SignalFormatter;
import io.github.oclay1st.wfdb.formatters.SignalFormatter61;
import io.github.oclay1st.wfdb.mocks.MockHeaderSignal;

class SignalFormatter61Test {

    @Test
    @DisplayName("Should convert raw data to signal samples with format 61")
    void shouldConvertFromRawDataToFormat61() {
        byte[] source = { 4, 3, 2, 1 };
        HeaderSignal signal = new MockHeaderSignal.Builder()
                .format(SignalFormat.FORMAT_61)
                .initialValue(1027)
                .build();
        HeaderSignal[] headerSignals = { signal };
        SignalFormatter formatter = new SignalFormatter61();
        int[] formattedSamples = formatter.convertBytesToSamples(source, headerSignals);
        assertNotNull(formattedSamples);
        assertArrayEquals(new int[] { 1027, 513 }, formattedSamples);
    }

    @Test
    @DisplayName("Should convert signal samples with format 61 to raw data")
    void shouldConvertFromFormat61ToRawData() {
        int[] samples = {1027, 513};
        HeaderSignal signal = new MockHeaderSignal.Builder()
                .format(SignalFormat.FORMAT_61)
                .initialValue(1027)
                .build();
        HeaderSignal[] headerSignals = { signal };
        SignalFormatter formatter = new SignalFormatter61();
        byte[] source = formatter.convertSamplesToBytes(samples, headerSignals);
        assertNotNull(source);
        assertArrayEquals(new byte[] { 4, 3, 2, 1  }, source);
    }

}
